+++ src/pages/games/AppleGame/game/view/appleGame/AppleGameView.tsx
