+++ src/pages/games/AppleGame/game/view/appleGame/ClassicMode.tsx
