+++ src/pages/games/AppleGame/game/view/appleGame/DefaultMode.tsx
