+++ src/pages/games/AppleGame/game/view/snackGame/SnackGameMode.tsx
