+++ src/pages/games/AppleGame/game/view/snackGame/SnackGameView.tsx
